+++ lib/statistics/providers.rb
@@ -5,3 +5,4 @@ end
 
 require_relative 'providers/connpass'
 require_relative 'providers/doorkeeper'
+require_relative 'providers/facebook'
