+++ spec/lib/statistics/client_spec.rb
@@ -2,18 +2,4 @@ require 'rails_helper'
 require 'statistics'
 
 RSpec.describe Statistics::Client do
-  context 'Facebook' do
-    include_context 'Use stubs for Facebook'
-
-    describe '#fetch_events' do
-      subject { Statistics::Client::Facebook.new.fetch_events(group_id: 123451234512345) }
-
-      it do
-        expect(subject).to be_instance_of(Array)
-        expect(subject.size).to eq 1
-        expect(subject.first['id']).to eq '125500978166443'
-        expect(subject.first.dig('owner', 'id')).to eq '123451234512345'
-      end
-    end
-  end
 end
